+++ registry/handlers/tags.go
@@ -117,7 +117,7 @@ func (th *tagsHandler) GetTags(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.Header().Set("Content-Type", "application/json")
 
 	// Add a link header if there are more entries to retrieve (only supported by the metadata database backend)
 	if moreEntries {